requirements.txt
requests>=2.28.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi
uvicorn
